+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/plan02/day01/Solution019_01.java
@@ -1,36 +1,33 @@
-package com.noob.algorithm.daily.plan02.day02;
+package com.noob.algorithm.daily.plan02.day01;
 
 import com.noob.algorithm.daily.base.ListNode;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Stack;
 
 /**
  * 🟡 019 删除链表的倒数第N个节点 - https://leetcode.cn/problems/remove-nth-node-from-end-of-list/description/
  */
-public class Solution019_02 {
+public class Solution019_01 {
 
     /**
-     * 思路分析：模拟法，借助栈辅助遍历（链表节点入栈）
-     * 栈思路：遍历节点入栈，随后弹出n个节点，第n+1个节点即为待删除项
+     * 思路分析：模拟法，获取到链表长度，然后找到 len - n - 1 （找到待删除节点的前一个节点）
+     * 例如：[1,2,3,4,5] n=2，删除倒数第2个节点得到[1,2,3,5],那么应该找到第3（索引为5-2-1，len-n-1）个节点执行删除操作
      */
     public ListNode removeNthFromEnd(ListNode head, int n) {
         // 构建虚拟头节点
         ListNode dummy = new ListNode(-1, head);
         // 定义遍历指针
         ListNode cur = dummy;
-        Stack<ListNode> stack = new Stack<>();
+        List<ListNode> list = new ArrayList<>();
         while (cur != null) {
-            stack.add(cur);
+            list.add(cur);
             cur = cur.next;
         }
 
         // 遍历节点
-        while (n-- > 0) {
-            stack.pop();
-        }
-        ListNode pre = stack.pop();
+        int len = list.size();
+        ListNode pre = list.get(len - n - 1);
         if (pre.next != null) {
             pre.next = pre.next.next;
         }
